Update patch set 4
Patch Set 4: Code-Review-1

Aswin, last changes look good. only comment is the todo's for the isLastPort stuff. Should we wait until that is tested before committing?

Other question, are these rules there by default? Meaning when I create a network with neutron will the sg rules for the dhcp be expected - meaning that now there is code for handling sg rules then the flows will be pushed? So that we need to make sure thsi stuff is workin?

Patch-set: 4
Label: Code-Review=-1