Update patch set 1
Patch Set 1:

flavio, looking good!

One quick comment, could we add the external-ids and other-config? other-config is good to see so we know the local_ip is set correctly. And for vlans that the provider_mappings are good

external-ids are good to know what neutron port is mapped. Which by the way, can your tool cross-reference this mdsal data with the neutron data?

Patch-set: 1