Update patch set 1
Patch Set 1:

(24016) add the collection in the README for stable/lithium.
(24015) add the collection in the README for master.

I had to push new patches because the one holding the postman collection in master has already been merged.

So Flavio, please +1 this one and the mentioned gerrit above will add the reference to this postamn in README as asked.

Also regarding the question I left in a previous comment, I saw this conversation so I guess it ok to have this collection in stable/lithium:

09:27]  <shague> regXboi: Ryan, are the urls for lbass the same in Li and Be?
[09:28]  <shague> when Alexis testsed against Be he  had to use controller/nb/v2/neutron/lbaas/* instead of controller/nb/v2/neutron/*
[09:29]  <regXboi> shage: in Li both /lbass and /* were supported for lbaas
[09:29]  <regXboi> shage: in Be lbaas goes back to lining up with openstack and so only /lbaas will be supported
[09:30]  <regXboi> IIRC, the classes in Li that supported /* should be marked deprecated
[09:30]  <regXboi> (at least I hope they are)
[09:30]  <regXboi> er
[09:30]  <regXboi> shague: ^^^^
[09:30]  * regXboi fingers aren't working this morning
[09:30]  <shague> regXboi: thanks. Ok cool. Alexis made changes in Be and I wanted to cherry-pick them to stable/li so sounds like I can to keep things consistent
[09:31]  <regXboi> yes, you should be able to

Patch-set: 1
Reviewer: Gerrit User 3675 <3675@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0