Sonar clean-up: OF13Provider
This class contains lots of methods with unused parameters, but they
follow a pattern so override that instead of cleaning up the
parameters.

Dead-code elimination (commented code, unused methods, unnecessary
initialisations).

getTunnelReadinessStatus() appears unused but is tested for in the
corresponding unit-test class, so keep it for now and override the
unused warning.

Long dpid -> long dpid.

if (!local) {} if (local) {} -> if (!local) {} else {} (the value of
local doesn't change in the first block).

networkingProviderManager is initialised and used in a single method,
so convert it to a local variable and adjust the corresponding test.

Return value simplification (? true : false isn't necessary).

<> operator in OF13ProviderTest.

Change-Id: Ibf74d332257f063ffc603e663b1707829ddaa6e3
Signed-off-by: Stephen Kitt <skitt@redhat.com>