Update patch set 1
Patch Set 1: Code-Review+2

nice, so that worked?

thanks for getting the netvirt-it also.

As a follow-on patch it would be better to remove the need to use the ovsdb and southbound karaf distros. The only reason we use those is because around the karaf 3.0.3 transition the it was broken if the integration distro was used along with loading the required features. I think it had to do with the karaf 3.0.3 being mroe strict on how it found the bundles or how they were included in the pom dependencies. that might be fixed and better understood now.

Patch-set: 1
Label: Code-Review=+2