Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

hmm... I kanda agree with Isaku in regards to the where this change is made. The netvirt code is actually ipv6 ready; just this provider code that it is not ipv6 capable. You can see that from the type of the parameters provided.

the only change I'd do is to change the success to a notSupported status.

Patch-set: 1
Label: Code-Review=-1