Update patch set 1
Patch Set 1:

Raksha,

your test seems to show that it works right and you can still update the original connection. But having a second config in the datastore seems like a problem waiting to happen.

Why didn't the second request get dropped? I thought the patch would drop extra requests. If not, then maybe that should be added. I am surprised that we actually get a second created dataChange, though I guess with the different nodeId it could happen. So maybe something should be added to drop any requests to the same node.

Would like Ed to weigh in on this.

Thanks, Sam

Patch-set: 1