Update patch set 1
Patch Set 1:

Yes - I think multiple config nodes do represent a problem. such as what happens when we add the functionality to reconnect with nodes after a controller restart - may not be the same config node that was active previously.

However, by the time the southbound gets the data change event, the node is already in the config datastore - right?  Is it possible to for the southbound to delete the item fromt he md-sal?  Is that a valid thing to do  if it is possible?

Patch-set: 1