Update patch set 1
Patch Set 1:

By default if you don't define it <relativePath> defines itself as ".." or the directory directly above itself. However this is not what we want for odlparent because the directory above it project directory is not where odlparent lives. By explicitly stating that <relativePath/> as nothing we are able to force Maven to skip the implicit behaviour and directly tell Maven to look online at nexus.opendaylight.org.

Additionally autorelease requires this field as it does a grep for these lines and replaces all parents with correct paths relative to autorelease's clones of all repos to produce the autorelease build so if we remove them autorelease will start failing to build.

Patch-set: 1