Fix for defect 4135
https://bugs.opendaylight.org/show_bug.cgi?id=4135

If integration bridge or external bridge exist prior to
connecting to ODL, then the pre-existing bridge is not added.
But the side effect is that the controller is not set on
this bridge.
Current code checks to see if the bridge already exists on the node. This
patch adds a check to see if the bridge exists on CONFIG datastore. If the
bridge does not exist on config, it indicates that the bridge existed prior
to ODL connection, and it adds the bridge.

Change-Id: Icc0718fb38296dd475f24dc4b2a5e97ecca00c71
Signed-off-by: Sharad Mishra <sharad.d.mishra@intel.com>