Update patch set 2
Patch Set 2:

Ok - thanks Anil.  I can give that a try.

After rebasing, the issue I'm up against is that this patch no longer causes the operational md-sal to get populated (monitoring does not get configured).

Ideally, if there was a way to cause the nodes to trigger a data change event again, that would cause the normal code flow to occur without replicating it for this case.  Any ideas on how to make that happened without actually deleting the entry and adding it back again?

Patch-set: 2