Update patch set 2
Patch Set 2:

Ideally whenever controller restart, clustering services replay all the data changes in config data store. So actually this method should be called whenever you controller reboots

https://github.com/opendaylight/ovsdb/blob/master/southbound/southbound-impl/src/main/java/org/opendaylight/ovsdb/southbound/OvsdbDataChangeListener.java#L70

I would suggest if you can dump some log in this method and see if you get any data from the data store when you reboot the controller. If we are getting the data and re-connection is not happening then probably an issue with this code, and if we are not getting the DCNs then we need to poke clustering team for that.

Patch-set: 2