Drop Karaf Pax Exam from default profile
Having two Pax Exam containers causes Pax Exam to fail at
initialisation; this is handled by disabling the Karaf container
altogether for failsafe-integration-tests, but it prevents
Maven-compliant IDEs from running the integration tests.

Since there's a separate profile for the Karaf container, and the
NetvirtIT instructions indicate that should be use, there doesn't seem
to be any point in keeping the Karaf Pax Exam container in the default
profile.

Change-Id: I071eacb77705fbeda8847851fb8e16e9edd2b176
Signed-off-by: Stephen Kitt <skitt@redhat.com>