Update patch set 1
Patch Set 1: Code-Review-1

Stephen,can you check if karaf settings are even needed in the integrationtest dir?

I think we don't even need any of the karaf container stuff here anymore. This might have been added  when we were first trying to get karaf based it working. Since then we moved the karaf based it into the Southbound and the NetVirt.

Which then begs the question on how we want to structure IT going forward. We had the idea to keep all the IT together since we thought it could share a parent pom. That seemed to make sense with the non-karaf stuff, but with karaf-based IT the parent came from config-parent. Any thoughts on the bets way to structure stuff going forward?

We also want to make sure sonar is reported properly. Currently it looks like southbound-it is not reporting sonar correctly.

Patch-set: 1
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1