Update patch set 1
Patch Set 1:

It seems that some work has been done here [1] to add an overlay on config-parent to have a specific IT-parent that seems to report correctly.
I know the startup archetype now creates an IT folder, so based on this, I think IT should resides in its related module: so southbound IT should be in southbound, netvirt IT should be in netvirt and so on.

[1] https://github.com/opendaylight/controller/blob/master/opendaylight/md-sal/mdsal-it-parent/pom.xml

Patch-set: 1
Reviewer: Gerrit User 3675 <3675@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0