Bug 4160 - null pointer exception in getDHCPServerPort() (redo)
null pointer exception in SecurityServicesImpl.getDHCPServerPort()
because neutronPortCache.getPort() may return null. It needs null check.
NOTE: the order of neutron port creation/deletion via neutron northbound
api and ovs port creation/deletion is not deterministic. Either order can
happen. SecurityServicesImpl seems to be fixed to be aware of it.
it's another story. This patch addresses only null pointer exception
for now.

This is a redo, to fix a merge mistake introduced in
https://git.opendaylight.org/gerrit/#/c/24579/8..9

Original-Change-Id: I40f78b1477f1c2a6ad3e23b2e60e21be6d9abc23
Change-Id: Ib929bf774426d6240934ea406189e25be1ecf536
Co-Authored-By: Isaku Yamahata <isaku.yamahata@intel.com>
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>