Update patch set 1
Patch Set 1:

As I ran into issues updating this patch, i created https://git.opendaylight.org/gerrit/#/c/26118/1. Please review this. As discussed in the ovsdb meeting, the patch prevents the addition of duplicate node with same connection info in the operational md-sal. It provides a warning message to the user and the user should take care of duplicate node id in the config md-sal.

Patch-set: 1