Update patch set 1
Patch Set 1: Code-Review+1

Seen how cluster/HA subject expands in different areas, I plan to start a cluster test effort and coordination across different projects and under the unbrella of integration/test project. I will send a mail later for this but this is one of the steps we need for verifying new ofplugin works with ovsdb code, also note this feature is only created for test purposes (at lest for now).

Patch-set: 1