Mdsal utils bundle
Patch set 2: Renamed to MdsalUtils and matched the strucure of
the other bundles in the utils dir.

This commit adds a Mdsal utils bundle that has all the mdsal methods to read,
write, delete and merge to the datastore. This is so that other bundles can
reuse the same methods.

Change-Id: Ida83ed957f74155224b007b54e3e80bfb1e29a71
Signed-off-by: Sam Hague <shague@redhat.com>