Create patch set 1
Patch-set: 1
Subject: Added manager list to Ovsdb Node. For application southbound plugin only give a device level view and not a connection level view. To make net-virt cluster aware, it needs to the managers of the ovsdb server, so that when it creates ovs bridges, it can connect them to the same controllers through OF plugin
Commit: 1da353f00419e47c00ca1f8010ab747737d195ee