Create change
Patch-set: 1
Change-id: I8b68fef9583f4fbcc330a0e9fc239015851b3f7c
Subject: Added manager list to Ovsdb Node. For application southbound plugin only give a device level view and not a connection level view. To make net-virt cluster aware, it needs to the managers of the ovsdb server, so that when it creates ovs bridges, it can co
Branch: refs/heads/topic/lithium/clustering