Update patch set 2
Patch Set 2: Code-Review+1

I can't make that change actually.  Although compelling, we can't reuse the ExecutorService.  In the instance that it is shutdown (long running test that we mark as a fail), the executor is shut down.  There is no way to restart it.  I tried this for about 3 hours yesterday, and couldn't get it to work.  I'd say this is good as it is, and ready for merge pending further feedback.

Patch-set: 2
Reviewer: Gerrit User 3635 <3635@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1