Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Ravi: thanks for finding the issue! Because of you I now understand
the bug. I have a suggestion in making the fix a bit better, so we
do not fragment the places where the router caches are updated. I'll
submit a patch set with them and ask for your feedback.

Patch-set: 1
Label: Code-Review=-1