4408 - Inter-network East-west traffic after unstack/stack
The static Arp and L3 forwarding rules for inter-network
east west traffic were not programmed after a unstack/restack
of a compute node
Registered with NodeCacheManagerImpl to receive the node
Add/DEL event. In the handler cleared the static Arp and
L3 forwarding rules on node deletion. When a new compute
node is added with a new VM, the flows will be programmed
properly
Change-Id: Ia736c3d16333b9c80891b707b161976e0011f003
Signed-off-by: ravi_sabapathy <ravi_sabapathy@dell.com>