Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: If controller joins the cluster late, and the owner of the device is already decided, EntityOwnershipService won't notify new candidate that register for ownership. Adding a proactive check on the ownership state of the device, so if it's already owned, adding the connection instances in the connection cache.
Commit: 91461a3aef97eb358ac0fe7d96b448c5f6ede0ca