Update patch set 1
Patch Set 1:

This patch looks fine to me, once all the comments form chris is addressed. But i want to share a thought here, As per my understanding Unicast ARP are not considered good in the network( e.g snort systems can raise alarm, proxy ARP don't work with it etc), but i am not sure whether this is considered bad in general, but i think it has a side effect that if somebody disabled it unicast arp on switch, arp resolution can fail. The intention behind using the unicast arp is to not create unnecessary broadcast traffic within host. I can suggest an alternate approach, if we agree that unicast arp can be problematic. We do arp resolution through br-ex bridge and this bridge has two ports only, one patch port to br-int and another port that connects to external network. This is the static configuration that we need to run the net-virt and it's all in our control. While sending the broadcast arp request packet, we can avoid all the internal ports (like patch port in this case) and send to the remaining port (which is only external port). So it's control broadcast of the broadcast packet. It can probably help us fix both the problem 1) less broadcast traffic 2) no problem because of unicast arp. thoughts?

Patch-set: 1
Label: Code-Review=0