Update patch set 1
Patch Set 1: Code-Review-1

Before I go back and address all of the individual comments, we should probably answer Anil's basic question about whether to make this change at all.  

Unicast ARPs are done, and are suggested by RFC 1122 as an ARP cache validation technique.  It seems like a reasonable thing for the controller to do to validate the MAC addresses.  Based on what I've read so far, we probably want to back off our ARP rate a bit, and if we keep the unicast MAC option, we may want to tolerate a couple missed replies.

However, when you Google "unicast ARP", the first thing you see is a question about why Snort has a rule to detect them, so it makes you wonder if it's a good thing.  If unicast ARP has become a security concern, we should at least have a way to turn it off.  I'm not sure how big a deal proxy ARP will be in this situation, but that's something else to think about.

I think that Anil's suggestion to just send ARPs on the external port on br-ex is a good one -- that's the only place we'll find the external gateway, so there's no point sending on the internal network.  However, I don't think it solves the broadcast issue because this port would typically be connected to a VLAN on the underlay network, and the ARPs would be broadcast to every compute node in the network (and anything else that happens to be connected to the network).

I'll need to look into this more, but would appreciate any opinions.

Patch-set: 1
Label: Code-Review=-1