Create patch set 1
Patch-set: 1
Subject: If controller joins the cluster late, and the owner of the device is already decided, EntityOwnershipService won't notify new candidate that register for ownership. Adding a proactive check on the ownership state of the device, so if it's already owned, adding the connection instances in the connection cache.
Commit: 077a73c3161a45859461289d3ebe60bdc0eaa670