Update patch set 2
Patch Set 2:

To Sam's point of finding the eth interface, i think we should refrain from doing that, because it's not necessary that public interface card are named like ethX, sys admin do use custom names as well. In my opinion we should skip the ports which we have control on, like the patch-port, which net-virt creates and that's the only port on this bridge except the public interface. So i would skip that port and out packet on rest of the port (which is basically public interface). Thoughts?

Patch-set: 2
Reviewer: Gerrit User 65 <65@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0