Update patch set 3
Patch Set 3:

can you give a little more info? My assumption is that when the node re-stacks
it will have all the mdsal flows re-programmed, so the cache in the ovsdb L3 netvirt is unaffected.

Clearing the cache doe not do anything about the programmed flows, thus I think I may be missing something here and gerrit [1] is not addressing the bug.

[1]:  https://git.opendaylight.org/gerrit/#/c/27946/3/

Patch-set: 3