Update patch set 3
Patch Set 3:

Okay... I got a better understanding on what is at play here.
The cache itself is not a problem, but id nodeId stays the same while the dpid is changed, then we would end up in a situation where openflowplugin and cache are out of sync.

I still think that the proposed fix is not the best approach to fixing this. A better approach would be to use 'dpid' as the key in the cache or eliminate the cache all together, assuming openflow plugin is idem-potent and fast enough to handle noops. Please let me know if this holds water. ;)

Patch-set: 3