Update patch set 3
Patch Set 3: Code-Review-1

With the current solution, NeutronL3Adapter will re-write the flows irrespective of whether the flows already exist or not. If flows don't exist, that should be fine, but if flows exist, that's unnecessary data store operation. I want to suggest an alternate solution-- whenever new node connects to the controller, check the configuration data store if dpid of the newly connected node exist, if yes, it's no-op, but if doesn't exist we can cleanup the cache. That will avoid unnecessary overwrite of the flows,given that it can be significant as most of the time connections flips rather then new node joins.

Patch-set: 3
Reviewer: Gerrit User 65 <65@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1