Update patch set 1
Patch Set 1:

> Andre,
 > One observation from my testing.
 > 1) Sometimes the iface (eth2) is down. So i had to explicitly turn
 > it up.
 > 

eth2 was up, and the ARPs from the GatewayMacResolver were being sent and received; however, the dest MAC in the OF rule mentioned was wrong.

 > 
 > The devstack script does an explicit iface up.  The br-ex creation
 > and port addition should also have been commented out in the
 > dev-stack scripts as we already discussed.
 > 

The following code from plugin.sh only creates br-ex and adds the port if $PUBLIC_INTERFACE is defined.  Since I'm not defining it, I don't think the lines inside the if statement will get executed, so I don't need to comment them out.

            # Add public interface to public bridge, if provided. Exclude 'br100' as a valid value for $PUBLIC_BRIDGE.
            if [[ -n "$PUBLIC_INTERFACE" && "$PUBLIC_INTERFACE" != "br100" ]]; then
                echo "Adding $PUBLIC_INTERFACE to $PUBLIC_BRIDGE"
                sudo ovs-vsctl add-port $PUBLIC_BRIDGE $PUBLIC_INTERFACE
                sudo ip link set $PUBLIC_INTERFACE up
            fi

 > Regards
 > Badri

Patch-set: 1