Update patch set 1
Patch Set 1:

Sam, so ODL_PROVIDER_MAPPING is set as a other config for OpenVSwitch table entry? and who sets it? is it devstack or OpenStack Nova? Because as far as i know local_ip is not set by anyone in the non-devstack deployment and admin need to set it explicitly. So if this very specific to devstack, then for non-devstack environment user will have to set it explicitly, is that correct? If this is the case then wondering which option will be more convenient for user, create a bridge and add whatever public port it wants to use, or set the mapping in OpenVswitch table and let controller take care of it.

Patch-set: 1