Update patch set 1
Patch Set 1:

Anil, yes the setting in something that needs to be populated. devstack uses the value to write to other_config. It is the same as for PROVIDER_MAPPINGS where the bridges are mapped to the physnet's. So the idea is just following previous examples of where there is config the user has to add in non-devstack environment and where the local.conf has flags to do that for devstack setups.

Patch-set: 1