Update patch set 4
Patch Set 4: Code-Review-1

(4 comments)

it looks to me that the cache is needed in SecurityServicesImpl.java
and that is the solely user of it. Move it out of the L3Adapter code.

If a callback from L3Adapter is needed to trigger the removal of L3
interface/router, SecurityServicesImpl should be able to get it straight out of neutron-northbound.

Patch-set: 4