Ensure extracted data isn't overwritten
Currently in extractCreatedOrUpdatedOrRemoved(), the value extracted
is always replaced by whetever was deleted, or null if nothing was
deleted. I fail to see how this can work; if a node is updated or
created, it will be added to the map with the appropriate key, then
removed (result.put(key, null) effectively). Instead of this, it seems
better to check that a non-null value is returned by getCreated(),
getUpdated() or getRemoved(), and only add the value to the map in
that case.

Change-Id: I31b0014b988451a04726dbdaa9da4481c750f166
Signed-off-by: Stephen Kitt <skitt@redhat.com>