Update patch set 1
Patch Set 1: Code-Review-1

Thanks for this gerrit, Sridhar!

I think there are a few more places that need to be visited in order to avoid this problem.
Ideally, we 1) should be consistent on the way we handle ipv6 and 2) try to centralize
that code into the flow programming. See [1] [2] for a gerrits on L3 that accomplished that.
In other words, if we could, we should move the checks to files in .../providers/openflow13/services/*.java

I also see other places [3] in the SG code that is assuming the address is IPv4. Can we extend the fixes to
those as well, under this gerrit?

Thanks again,

-- flavio

[1]: https://git.opendaylight.org/gerrit/#/c/26282/
[2]: https://git.opendaylight.org/gerrit/#/c/27846/2/openstack/net-virt-providers/src/main/java/org/opendaylight/ovsdb/openstack/netvirt/providers/openflow13/services/RoutingService.java
[3]: https://gist.github.com/a723f42504a21209b9fa

Patch-set: 1
Label: Code-Review=-1