Update patch set 1
Patch Set 1:

Gabriel,
The issue I ran into is that rows in the QoS and Queue tables can exist independent of being connected with a Port (they are root tables).  

So, for example, one could create a QoS entry on the host (say using ovs-vsctl).  When ODL receives the update, since it is not yet associated with a Port, it's not clear what could be updated in the model.  If that QoS entry was subsequently associated with a Port (again with ovs-vsctl), the UUID of the QoS entry is provided in the Port update, but the details of the QoS entry are not (they came earlier).

Patch-set: 1
Reviewer: Gerrit User 856 <856@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0