Fix updatedUMacsRemoteRows null check
The null check confuses Sonar; since the member variables are never
null (TyperUtils.extractRowsUpdated never returns null), and are only
initialised in the constructor, make them final and drop the null
check. The isEmpty() check is redundant with the for loop.

Remove the logger which isn't used.

Clean up some whitespace issues.

Change-Id: Ia234546957ff11901300149a5bac45c7bc0c9c96
Signed-off-by: Stephen Kitt <skitt@redhat.com>