Update patch set 1
Patch Set 1: Code-Review+1

The code looks good.  Have we seen this error?

When resolveMacAddress is called, it looks like the IP address should always get added to gatewayToArpMetadataMap before the first ARP is sent.

However, if not, then it should be safe to ignore because it should be there before the next one.

If this happens after the stopPeriodicRefresh has been called, it should be okay to ignore since we don't care about it anymore.

Another possibility is that we are receiving gratuitous ARP responses from IP addresses we don't care about.  These should be fine to ignore too.

If we can convince ourselves that this is the result of one of these expected cases, the we shouldn't even need the warning log.

Patch-set: 1
Reviewer: Gerrit User 3972 <3972@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1