Update patch set 2
Patch Set 2:

> Flavio, I see the ssh and http SG flows, but I didn't see any
 > traffic hitting them. Is that a concern?

not really, I just did not make any traffic on the ports I added to SG. HOWEVER... notice that the default rule for the tables is still a 'goto' instead of drop. That makes the SG rules not block any other traffic and that is intentional until SG implementation is declared mature by Aswin et all. That is the case on Beryllium as well, btw.
@Aswin, is that correct?

Patch-set: 2