Update patch set 23
Patch Set 23:

> what will happen with the ovssfc piece?

Gald you mentioned that. Since the beginning this piece was giving me trouble because the module wasn't there is my eclipse so I was not understanding the reason of the failure.

Regardless, this code is useless and is never used. 
Previous patch removed the feature that "seemed" associated to this piece of code but in fact the feature wasn't even pulling one bundle.

So regarding the code under ovssfc, up to you: do you want me to remove it?

Patch-set: 23
Label: Verified=0