Update patch set 23
Patch Set 23: Code-Review-1

> > I'd go for odlparent-lite where possible.
 > 
 > Stephen, I agree! The thing is maven-deploy-plugin and
 > maven-install-plugin are provided by odlparent and not by
 > odlparent-lite. As aggregator uses those plugin, instead of
 > re-defining the version, we rather liked using odlparent

The intention is that POMs which shouldn't deploy artifacts inherit from odlparent-lite (which isn't perfect yet). It's also a good idea to skip deploy/install, and Maven's warnings about the lack of versioning are OK to ignore since all we're doing is skipping the plugins (so any behaviour changes won't have an impact). Please switch to odlparent-lite.

Patch-set: 23
Label: Code-Review=-1