Update patch set 3
Patch Set 3:

Anil,

I changed the subject of this gerrit to reflect that it is for getting the Southbound to only monitor for the Open_vSwitch table. That is independent of the shared connection and needed to be fixed.

I sent another email discussing the shared connection stuff.

Thanks, Sam

Patch-set: 3