Update patch set 1
Patch Set 1:

This is an alternative to https://git.opendaylight.org/gerrit/#/c/31899/.

I discussed the netvirt features with Sam this morning, and we thought that it would be better to add the netvirt features to the top level ovsdb features file and do away with the netvirt-specific features.

Also, we don't think any netvirt-specific impl code will be needed -- the impl is in the renderers.  So, I deleted netvirt/impl.  This clears up the naming conflict.  If we decide we need an impl later, we can deal with it then.

Patch-set: 1
Label: Verified=0