Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

As noted in the review comments also, removing the feature might break the IT. I need to run some tests to make sure the IT still executes properly.

Why do we need to remove the auto load of the feature? This is just in the ovsdb netvirt karaf local distro. To test clustering you simply remove the entry from the boot features in the etc file.

Patch-set: 2