Update patch set 2
Patch Set 2:

Whoops, hit +2 also..

Are you saying changing or removing the featuresBoot line from etc/org.apache.karaf.features.cfg is an issue?

The IT may still work. The karaf pom has the line to add the default boot feature. In the IT config there is a getFeatureName() method also. We currently use both. I think we just need to verify that not setting it in the pom and then setting it in the IT still works.

Patch-set: 2