Add support for table offset
This commit adds a new config parameter table-offset to
allow setting the openflow table base. The default value is 0.
The user can change the value via the config file
netvirt-providers-impl-default-config.xml or by using the restconf
api show below. The value used must be between 1 and 145 - with 145
coming from 255-110 - 110 is the max table in the netvirt pipeline.

The following curl commands can be used to configure the value:
- Get the current value:
curl -u admin:admin http://localhost:8181/restconf/config/netvirt-providers-config:netvirt-providers-config
- Change the value: In this example the value is changed to 1
curl -u admin:admin http://localhost:8181/restconf/config/netvirt-providers-config:netvirt-providers-config -i -H "Content-Type: application/json" --data '{"netvirt-providers-config": {"table-offset": 1}}' -X PUT

If the table-offset is set, then an extra flow is added to the br-ex and br-int
switches to steer packets from table 0 to the first table in the pipeline.

IT tests were all updated to check for the new table-offsets as well as
adding unti-tests. Some IT tests have a little more due to the Yang
pacth that enforces strictre rules on the Yang files and was failing the tests.

Further patches should add config service in netvirt to be a central point
handling all the config updates. SG is also adding a config value that should
be included.

Change-Id: I44b3a442e8050d8f1ef1c5cd9db5c32af2d5ce19
Signed-off-by: Sam Hague <shague@redhat.com>