Update patch set 1
Patch Set 1:

Sam, this is very specific to net-virt code, where it checks whether *this* instance of net-virt is master instance or not.It's moreover a wrapper around the utils.mdsal-utils, which calls the mdsalutils method if it's master instance. Each application will have to decide how they want to define their behavior in case of clustering. So if we want something common for each application in net-virt, we need to put this at some shareable location where each module can access it, and that's going to create a dependency between these modules and i am not sure we want to go that way.

Patch-set: 1
Reviewer: Gerrit User 65 <65@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0