Update patch set 1
Patch Set 1:

Did you attend the statistic & status issue in some other way?
I have not touch this code in a while, I can revisit the code and see if it scales. 

As i remember the issue is that the statistics & status are constantly polled so you need to extract them from the main query and move them to a query of their own so if the OVS has more than 50 bridges +ports, the ovsdb won't send you all the data every second, just the statistics & the status one.
BTW, those are the same statistics as the openflow ones, is there any need for duplicity? if so, i recommend putting some config to disable statistic polling when it is polled from the openflow...

Patch-set: 1
Label: Verified=0