Update patch set 1
Patch Set 1:

We have not implemented any other solution, so I was looking back at this patch to see if you had done any more verification with it. Some older using code relied on these extra ovsdb updates to do some work but a lot of that has been refactored with the latest mdsal architecture so I don't think we need the frequent updates. Noone uses stats or status from the ovsdb for openflow so that is another possible reason to lower the refresh. Ovsdb also allows the controller to increase the polling rate so that is another option.

Patch-set: 1