Update patch set 3
Patch Set 3:

Hello Sam,

As you know, for a VM to configure an IPv6 address it needs the Router Advts (for all the modes SLAAC/DHCPv6 Stateless and Stateful), so this feature will be kind of useful mainly when using Neutron L3 agent. 
While using ODL_L3 (i.e., L3 support from ODL) since we are currently not supporting IPv6 router, VMs will not be able to configure an IPv6 address (so IPv6 SGs may not be of much help in this scenario).
Just wanted to share this info, so that we are on the same page.

Now coming back to this patch. I verified the changes on the master branch and not yet on stable/lithium. I can verify the changes on the stable/lithium branch and let you know. 
IMHO this is good for master branch (as we have plans to support IPv6 router in Boron), but I'm not sure of backporting to stable/lithium (for the above reasons). 
I'll be interested to know Flavio's comments.

Thanks,
--Sridhar.

Patch-set: 3