Rewrite NeutronL3AdapterTest without PowerMock [WIP]
This patch builds the tests on top of a mock OSGi context managed by
Apache Sling, and uses AbstractDataBrokerTest to provide the MD-SAL
layer. This allows the tests to stick much closer to the environment
which actually ends up being used, and makes sure the initialisation
is as used in production.

Currently most of the tests are broken, I wanted to get the setup out
there for comment.

Change-Id: If1afaa0b5cb388dac8cfb620fc9e68a3348526bb
Signed-off-by: Stephen Kitt <skitt@redhat.com>