BUG-4917 Ensure all OVS flows are deleted
This commit is basically a revert of this change:
https://git.opendaylight.org/gerrit/#/c/27521/
With that change the cache cleanup in fact never ran because the
port is removed before the code is actually run.
The bug for which that commit was a fix is resolved
with this code new commit as well.

In addition, updateL3ForNeutronPort() now allows
deleting flows even when networkId is not found
in networkIdToRouterMacCache. This takes care of
some of the DHCP issues.

Change-Id: Idc3c1ff4b57f5471615ccc6c627ca8f411c37db2
Signed-off-by: Josh <jhershbe@redhat.com>