BUG 5164 Cleanup Floating IP Rules when VM is deleted.
Patch set 2: Fixed unit test.

When a VM is deleted, Neutron does not send a notification
to update/delete any floatingip that may have been associated
with a port. As a result, openflow rules for the floating ip
are left installed.

Neutron team says this is expected behavior for floating ip.
So, to cleanup, use neutron port delete event, and find
any associated floating ip, and cleanup installed OF
rules.

Change-Id: If9e2ffc5b16f0da636b3e9b6b1c4fbfd46fd6aff
Signed-off-by: Victor Pickard <vpickard@redhat.com>